move scanner/models to models
