+++ depobs/clients/cratesio.py
@@ -5,7 +5,7 @@ from typing import AbstractSet, Dict, AsyncGenerator, Generator, Optional
 import aiohttp
 
 from depobs.clients.aiohttp_client_config import AIOHTTPClientConfig
-from depobs.scanner.models.rust import (
+from depobs.models.rust import (
     RustPackageID,
     cargo_metadata_to_rust_crates,
 )
