+++ depobs/worker/serializers.py
@@ -19,7 +19,7 @@ from typing import (
 
 from depobs.database.models import Advisory, NPMRegistryEntry, NPMSIOScore
 from depobs.util.graph_util import npm_packages_to_networkx_digraph, get_graph_stats
-from depobs.scanner.models.nodejs import NPMPackage, flatten_deps
+from depobs.models.nodejs import NPMPackage, flatten_deps
 from depobs.util.serialize_util import (
     get_in,
     extract_fields,
