+++ tests/fixtures/graphql/REPO_DEP_MANIFESTS_first_selection.graphql
