+++ tests/fixtures/graphql/REPO_DEP_MANIFESTS_next_selection.graphql
