+++ tests/fixtures/graphql/REPO_DEP_MANIFEST_DEPS_first_selection.graphql
