+++ tests/fixtures/graphql/REPO_DEP_MANIFEST_DEPS_nested_first_selection.graphql
