+++ tests/fixtures/graphql/REPO_DEP_MANIFEST_DEPS_next_selection.graphql
