+++ tests/fixtures/graphql/REPO_LANGS_first_selection.graphql
