+++ tests/fixtures/graphql/REPO_LANGS_next_selection.graphql
