+++ tests/fixtures/graphql/REPO_VULN_ALERTS_first_selection.graphql
