+++ tests/fixtures/graphql/REPO_VULN_ALERTS_next_selection.graphql
