+++ tests/fixtures/graphql/REPO_VULN_ALERT_VULNS_first_selection.graphql
