+++ tests/fixtures/graphql/REPO_VULN_ALERT_VULNS_nested_first_selection.graphql
