+++ tests/fixtures/graphql/REPO_VULN_ALERT_VULNS_next_selection.graphql
