+++ tests/fixtures/graphql/REPO_first_selection.graphql
