+++ tests/fixtures/nodejs/flatten/input/nested_many_levels_with_decamelize_1.2.0_duplicate.json
