+++ tests/fixtures/nodejs/flatten/input/one_dep.json
