+++ tests/fixtures/nodejs/flatten/input/three_deps_two_nested.json
