+++ tests/fixtures/nodejs/flatten/input/two_deps.json
