+++ tests/fixtures/nodejs/flatten/input/zero_deps.json
