+++ tests/fixtures/nodejs/flatten/output/one_dep.json
