+++ tests/fixtures/nodejs/flatten/output/three_deps_two_nested.json
