+++ tests/fixtures/nodejs/flatten/output/two_deps.json
