+++ tests/fixtures/nodejs/flatten/output/zero_deps.json
