+++ tests/fixtures/nodejs/visit/input/zero_deps.json
