+++ tests/fixtures/nodejs/visit/output/three_deps_two_nested.json
