+++ tests/fixtures/nodejs/visit/output/zero_deps.json
